+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## [1.0.0+69]
 
-- cache düzenlemeleri.
+- Add gallery and story management features to UserAccounts model; update version to 1.0.0+70.